+++ rpc/namespaces/ethereum/net/api.go
@@ -19,9 +19,9 @@ import (
 	"context"
 	"fmt"
 
+	rpcclient "github.com/cometbft/cometbft/rpc/client"
 	"github.com/cosmos/cosmos-sdk/client"
 	ethermint "github.com/evmos/ethermint/types"
-	rpcclient "github.com/tendermint/tendermint/rpc/client"
 )
 
 // PublicAPI is the eth_ prefixed set of APIs in the Web3 JSON-RPC spec.
@@ -38,9 +38,15 @@ func NewPublicAPI(clientCtx client.Context) *PublicAPI {
 		panic(err)
 	}
 
+	// FIXME: tmp solution to make code pass compilation
+	tmClient, ok := clientCtx.Client.(rpcclient.Client)
+	if !ok {
+		panic("client doesn't implement rpcclient.Client")
+	}
+
 	return &PublicAPI{
 		networkVersion: chainIDEpoch.Uint64(),
-		tmClient:       clientCtx.Client,
+		tmClient:       tmClient,
 	}
 }
 
@@ -50,6 +56,7 @@ func (s *PublicAPI) Version() string {
 }
 
 // Listening returns if client is actively listening for network connections.
+// DEPRECATED： client context doesn't implement NetInfo
 func (s *PublicAPI) Listening() bool {
 	ctx := context.Background()
 	netInfo, err := s.tmClient.NetInfo(ctx)
@@ -60,6 +67,7 @@ func (s *PublicAPI) Listening() bool {
 }
 
 // PeerCount returns the number of peers currently connected to the client.
+// DEPRECATED： client context doesn't implement NetInfo
 func (s *PublicAPI) PeerCount() int {
 	ctx := context.Background()
 	netInfo, err := s.tmClient.NetInfo(ctx)